+++ JCTO.Domain/Dtos/EntryDto.cs
@@ -8,7 +8,7 @@ namespace JCTO.Domain.Dtos
         public Guid CustomerId { get; set; }
         public Guid ProductId { get; set; }
         public string EntryNo { get; set; }
-        public double InitialQuantity { get; set; }
+        public decimal InitialQuantity { get; set; }
         public DateTime EntryDate { get; set; }
         public EntryStatus Status { get; set; }
     }
