+++ JCTO.Domain/Dtos/EntryListItemDto.cs
@@ -6,8 +6,8 @@ namespace JCTO.Domain.Dtos
     {
         public Guid Id { get; set; }
         public string EntryNo { get; set; }
-        public double InitialQuantity { get; set; }
-        public double RemainingQuantity { get; set; }
+        public decimal InitialQuantity { get; set; }
+        public decimal RemainingQuantity { get; set; }
         public string Product { get; set; }
         public DateTime EntryDate { get; set; }
         public string Customer { get; set; }
