+++ JCTO.Domain/Dtos/OrderListItemDto.cs
@@ -7,8 +7,8 @@ namespace JCTO.Domain.Dtos
         public Guid Id { get; set; }
         public string Customer { get; set; }
         public string Product { get; set; }
-        public double Quantity { get; set; }
-        public double? DeliveredQuantity { get; set; }
+        public decimal Quantity { get; set; }
+        public decimal? DeliveredQuantity { get; set; }
         public DateTime OrderDate { get; set; }
         public int OrderNo { get; set; }
         public string Buyer { get; set; }
